deliver self sourcecode and license to everyone requesting it
the license of sendhttp is the AGPL, so everyone
who modifies it must deliver the sourcecode to
everyone using his version of sendhttp. If
sendhttp always delivers it's own sourcecode,
no one must worry about not conforming to the
license. The only current drawback is, you can't
deliver regular files named "?source" or
"?license" anymore.